Fix a crash with DVD subtitles on x64 builds with software decoding.
Some code was forgotten in 0c8c08104c3e0d92529a01701260254704740241.

+++ decoder/LAVVideo/subtitles/LAVSubtitleConsumer.cpp
@@ -106,7 +106,7 @@ STDMETHODIMP CLAVSubtitleConsumer::ProcessFrame(LAVFrame *pFrame)
     }
 
     BYTE *data[4] = {0};
-    int stride[4] = {0};
+    ptrdiff_t stride[4] = {0};
     LAVPixelFormat format = pFrame->format;
     int bpp = pFrame->bpp;
 
@@ -286,7 +286,7 @@ STDMETHODIMP CLAVSubtitleConsumer::SelectBlendFunction()
   return S_OK;
 }
 
-STDMETHODIMP CLAVSubtitleConsumer::ProcessSubtitleBitmap(LAVPixelFormat pixFmt, int bpp, RECT videoRect, BYTE *videoData[4], int videoStride[4], RECT subRect, POINT subPosition, SIZE subSize, const uint8_t *rgbData, int pitch)
+STDMETHODIMP CLAVSubtitleConsumer::ProcessSubtitleBitmap(LAVPixelFormat pixFmt, int bpp, RECT videoRect, BYTE *videoData[4], ptrdiff_t videoStride[4], RECT subRect, POINT subPosition, SIZE subSize, const uint8_t *rgbData, int pitch)
 {
   if (subRect.left != 0 || subRect.top != 0) {
     DbgLog((LOG_ERROR, 10, L"ProcessSubtitleBitmap(): Left/Top in SubRect non-zero"));
